Dual Cards fix, ship position manager's code is moved to Setup SubPhase
