--- Assets/Scripts/ShipManagement/ShipPositionManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 0387792d8352295488c219acd5f0ece8
-timeCreated: 1494199065
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
